+++ src/js/api.js
@@ -137,19 +137,16 @@ export async function loadPackagesFromImports(
   if (imports.length === 0) {
     return;
   }
-  let packageNames = Module.packages.import_name_to_package_name;
+
+  let packageNames = Module._import_name_to_package_name;
   let packages = new Set();
   for (let name of imports) {
-    if (name in packageNames) {
-      packages.add(packageNames[name]);
+    if (packageNames.has(name)) {
+      packages.add(packageNames.get(name));
     }
   }
   if (packages.size) {
-    await loadPackage(
-      Array.from(packages.keys()),
-      messageCallback,
-      errorCallback
-    );
+    await loadPackage(Array.from(packages), messageCallback, errorCallback);
   }
 }
 
@@ -182,7 +179,7 @@ export function pyimport(name) {
  *        response = await fetch("./packages.json")
  *        packages = await response.json()
  *        # If final statement is an expression, its value is returned to
- * Javascript len(packages.dependencies.object_keys())
+ * Javascript len(packages['packages'].object_keys())
  *    `);
  *    console.log(result); // 72
  *